Added deterministic rng to region generation. Still need to replace all other instances of random number calls with the local Rng

+++ Dwarf/Helpers/pcg_basic.cpp
@@ -59,11 +59,11 @@ void pcg32_srandom(uint64_t seed, uint64_t seq)
 
 uint32_t pcg32_random_r(pcg32_random_t* rng)
 {
-    uint64_t oldstate = rng->state;
-    rng->state = oldstate * 6364136223846793005ULL + rng->inc;
-    uint32_t xorshifted = ((oldstate >> 18u) ^ oldstate) >> 27u;
-    uint32_t rot = oldstate >> 59u;
-    return (xorshifted >> rot) | (xorshifted << ((-rot) & 31));
+	uint64_t oldstate = rng->state;
+	rng->state = oldstate * 6364136223846793005ULL + rng->inc;
+	uint32_t xorshifted = ((oldstate >> 18u) ^ oldstate) >> 27u;
+	uint32_t rot = oldstate >> 59u;
+	return (uint32_t)((xorshifted >> rot) | (xorshifted << ((-rot) & 31)));
 }
 
 uint32_t pcg32_random()
